backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-socketio
psycopg2-binary
openai
python-docx
cryptography
pyopenssl
werkzeug